requirements.txt
python-dotenv==1.0.0
sanic==23.6.0
rich==13.7.0
httpx==0.26.0
typer==0.9.0
openai==0.27.8
emoji==2.7.0
